chore: remove conflicting /app directory for Next.js pages router deployment
